Refactor email service and update configuration.
Replaced `MailService` with `EmailService` for improved flexibility in sending templated emails. Updated environment and application configurations to support new API and local development profiles. Removed unnecessary CORS origin and added missing `.env` configuration for local Docker setup.
