+++ .env.local
@@ -1,0 +1,35 @@
+# Postgres
+POSTGRES_DB=project-place
+POSTGRES_USER=project-place
+POSTGRES_PASSWORD=project-place
+DB_URL=jdbc:postgresql://trackme-postgres:5432/project-place
+DB_USERNAME=project-place
+DB_PASSWORD=project-place
+
+# Redis
+REDIS_HOST=trackme-redis
+REDIS_PORT=6379
+REDIS_USER=trackme-redis
+REDIS_PASSWORD=trackme-redis
+REDIS_USER_PASSWORD=trackme-redis-pass
+
+# Backend
+BACKEND_DB_SCHEMA=backend
+
+# SSO и внутренние URL (остальное по необходимости)
+BACKEND_URI=http://host.docker.internal:8081
+SSO_URI=http://host.docker.internal:9000
+FRONTEND_URI=http://localhost:3000
+AFTER_LOGOUT_URI=http://localhost:3000
+AFTER_LOGIN_URL=http://localhost:3000/after-login
+REACT_APP_BACKEND_URI=http://127.0.0.1:8081
+
+# Frontend
+NODE_ENV=development
+REACT_APP_ENV=dev
+
+TRACKME_CLIENT_SECRET=password
+TRACKME_CLIENT_ID=track-me-client
+
+# Доп. переменные, если нужно:
+APP_NAME=track-me
