+++ docker-compose.yaml
@@ -113,8 +113,10 @@ services:
       - trackme-backend
       - trackme-sso
     environment:
-      - CORS_ORIGINS=http://localhost:3000,${FRONTEND_URI},${BACKEND_URI}
       - SPRING_PROFILES_ACTIVE=docker-local
+      - SERVER_REACTIVE_SESSION_COOKIE_SECURE=true
+      - SERVER_REACTIVE_SESSION_COOKIE_SAME_SITE=None
+
     env_file:
       - .env
     networks:
