+++ services/trackme-sso/src/main/java/net/akarmanov/projectplace/sso/services/EmailService.java
@@ -1,0 +1,12 @@
+package net.akarmanov.projectplace.sso.services;
+
+import java.util.Map;
+
+public interface EmailService {
+
+  void sendMail(String emailTo,
+                String emailForm,
+                String subject,
+                String templateName,
+                Map<String, Object> variables);
+}
