+++ services/trackme-sso/src/main/resources/application-security.yaml
@@ -31,7 +31,7 @@ spring:
                 - http://localhost:3000
               redirect-uris:
                 - ${BACKEND_URI:http://127.0.0.1:8081}/login/oauth2/code/track-me-client
-                - http://localhost:8081/sso/login/oauth2/code/track-me-client
+                - http://127.0.0.1:8081/login/oauth2/code/track-me-client
       resourceserver:
         jwt:
           issuer-uri: ${SSO_URI:http://localhost:9000}
